==> 1-intro.Rmd <==
---
title: "1. Intro to R"
subtitle: "Analytics Sandbox"
author: "K. Bret Staudt Willet | Florida State University"
date: "January 11, 2023"
output:
  xaringan::moon_reader:
    lib_dir: libs
    seal: true
    css: [default, 'css/custom.css', 'css/custom-fonts.css']
    nature:
      titleSlideClass: [left, bottom]
      highlightStyle: github
      highlightLines: true
      countIncrementalSlides: false
---

```{r setup, include=FALSE}
options(htmltools.dir.version = FALSE)

library(knitr)
library(tidyverse)
library(xaringan)
library(fontawesome)
```

class: inverse, center, middle

# `r fa("fas fa-otter", fill = "#fff")` <br><br> **Part 1:** <br> Introduction to R

---

# `r fa("fas fa-otter", fill = "#fff")` Our Constructivist Approach

--

1. We'll introduce some concepts
1. You'll try some code
1. We'll all discuss together

--

```{r, out.width = "360px", echo = FALSE, fig.align = "center"}
include_graphics("img/tech_support_cheat_sheet.png")
```

<div class="caption">
<p><a href="https://xkcd.com/627/" target="_blank">'Tech Support Cheat Sheet' from XKCD</a></p>
</div>

---

class: inverse, center, middle

# `r fa("fab fa-r-project", fill = "#fff")` <br><br> Background <br> on R and RStudio

---

# `r fa("fab fa-r-project", fill = "#fff")` Why Learn R?

--

- It is increasingly used in education

--

- It is cross-platform, open-source, and freely-available

--

- It is a programming language and quite flexible

--

- It is capable of carrying out basic and complex statistical analyses

--

- It is able to work with data small (*n* = 10) and large (*n* = 1,000,000+) efficiently

--

- There is a great, inclusive community of users and developers

---

# `r fa("fab fa-r-project", fill = "#fff")` Setting up R

### To download R

- Visit [**cran.r-project.org**](https://cran.r-project.org/) to download R
- Find your operating system (Mac, Windows, or Linux)
- Download the 'latest release' on the page for your operating system and download and install the application

### To download RStudio

- Visit [**rstudio.com**](https://rstudio.com/products/rstudio/download/) to download RStudio
- Find your operating system (Mac, Windows, or Linux)
- Download the 'latest release' on the page for your operating system and download and install the application

---

class: inverse, center, middle

# `r fa("fas fa-otter", fill = "#fff")` <br><br> Try it Out! 

---

# `r fa("fab fa-r-project", fill = "#fff")` Getting Started with RStudio

Activities:  

1. Running a single RMarkdown chunk  
1. Running another RMarkdown chunk  
1. Rendering an RMarkdown document to a PDF  
1. Creating your first visualization

---

# `r fa("fab fa-r-project", fill = "#fff")` RMarkdown

- RMarkdown is a data analysis "notebook" that combines text with code and output
- It is a great file type to use when beginning to use R and to create reproducible analyses
- It is fun to use because you can generate different types of output (Word, PDF, and even web-based)

---

class: inverse, center, middle

# `r fa("fas fa-otter", fill = "#fff")` Try it Out!

---

# `r fa("fas fa-otter", fill = "#fff")` Try it Out!

What do you think this code will do?

```{r, eval = FALSE}
starwars %>% 
  filter(sex == "female") %>% 
  select(name, hair_color, skin_color, homeworld)
```

---

# `r fa("fas fa-otter", fill = "#fff")` Try it Out!

Let's see!

```{r, eval = TRUE}
starwars %>% 
  filter(sex == "female") %>% 
  select(name, hair_color, skin_color, homeworld)
```

---

# `r fa("fas fa-otter", fill = "#fff")` Try it Out!

What do you think this code will do?

```{r, eval = FALSE}
starwars %>% 
  filter(sex %in% c("male", "none"),
         height <= 150) %>% 
  select(name, sex, height, mass, homeworld) %>%
  arrange(desc(height))
```

---

# `r fa("fas fa-otter", fill = "#fff")` Try it Out!

Let's see!

```{r, eval = TRUE}
starwars %>% 
  filter(sex %in% c("male", "none"),
         height <= 150) %>% 
  arrange(height) %>%
  select(name, sex, height, mass, homeworld)
```

---

# `r fa("fas fa-otter", fill = "#fff")` Try it Out!

```{r, echo = TRUE}
starwars %>% 
  unnest(starships) %>% 
  select(name, gender, starships) %>%
  head(10)
```

---

# `r fa("fas fa-otter", fill = "#fff")` Try it Out!

```{r, echo = TRUE}
starwars %>% 
  unnest(starships) %>% 
  mutate(vehicles = strsplit(starships, ",")) %>%
  unnest(starships) %>% 
  select(name, gender, starships) %>% 
  group_by(gender) %>% 
  count()
```

---

# `r fa("fas fa-otter", fill = "#fff")` Exploring further

```{r, echo = TRUE}
glimpse(starwars)
```

---

# `r fa("fas fa-otter", fill = "#fff")` Visualizing data

```{r, echo = TRUE, warning = FALSE}
starwars %>% 
  ggplot() +
  geom_point(aes(x = mass, y = height, color = gender),
             alpha = 0.5
             ) +
  theme_bw()
```

---

class: inverse, center, middle

# `r fa("fas fa-code", fill = "#fff")` <br><br> Try it out!

Hop over to [**Workbook 1**](workbooks/workbook1.Rmd)

---

class: inverse, center, middle

# `r fa("fas fa-list", fill = "#fff")` <br><br> Appendix: <br> Helpful Resources <br> and Troubleshooting

---

# Resources

```{r child="notes/resources.Rmd"} 
```

---

# Troubleshooting

```{r child="notes/troubleshooting.Rmd"} 
```
